requirements.txt
typer>=0.7.0
python-dotenv>=1.0.0
openai>=1.12.0
sentence-transformers>=2.2.2
scikit-learn>=1.3.0
pandas>=2.0.0
numpy>=1.25.0
spacy>=3.5.0
joblib>=1.3.0

# For testing
pytest>=8.0.0